+++ js/data.js
@@ -1,0 +1,66 @@
+const names = [
+  'Александр',
+  'Анна',
+  'Борис',
+  'Виктория',
+  'Владимир',
+  'Галина',
+  'Дмитрий',
+  'Евгения',
+  'Екатерина',
+  'Иван',
+  'Ирина',
+  'Константин',
+  'Лариса',
+  'Леонид',
+  'Маргарита',
+  'Михаил',
+  'Наталья',
+  'Николай',
+  'Ксения',
+  'Павел',
+  'Светлана',
+  'Сергей',
+  'Татьяна',
+  'Юрий',
+  'Яна',
+];
+
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const descriptions = [
+  'Задумчивый взгляд на фоне потрясающего заката.',
+  'Смех и радость на празднике.',
+  'Оживленная городская улица с высотными зданиями.',
+  'Дружеский пикник на природе с ароматным шашлыком.',
+  'Загадочная улыбка и блеск в глазах у прекрасной девушки.',
+  'Спокойствие и умиротворение: озеро в окружении гор.',
+  'Радость победы: спортсмен на пьедестале почета.',
+  'Уютная атмосфера домашнего вечера у камина.',
+  'Экстремальное развлечение: прыжок с парашютом.',
+  'Яркие краски осеннего леса: золотые листья и красные ягоды.',
+  'Прогулка на лошадях по зеленому полю: гармония с природой.',
+  'Неожиданный кадр: забавное выражение лица кота.',
+  'Романтический ужин на крыше под звездным небом.',
+  'Бескрайние просторы моря и одинокий парусник на горизонте.',
+  'Летний дождь: свежесть и чистота после грозы.',
+  'Творческий процесс создания картины: художник и его муза.',
+  'Уличные музыканты, играющие веселую мелодию для прохожих.',
+  'Зимние развлечения: веселые горки и радостные крики детей.',
+  'Завораживающий вид на ночной город с высоты птичьего полета.',
+  'Заснеженные деревья в зимнем лесу, окутанные лунным светом.',
+  'Девушка на фоне заката, танцует с длинными лентами в руках.',
+  'Стая птиц в полете над рекой, отражаясь в зеркальной поверхности.',
+  'Крупный план цветка розы с каплями росы на нежных лепестках.',
+  'Рыжий кот на подоконнике, наблюдающий за падающим снегом.',
+  'Разноцветные листья, собранные в осенний букет, на деревянном столе.',
+];
+
+export {names, messages, descriptions};
