+++ js/settings.js
@@ -1,0 +1,10 @@
+const settings = {
+  MAX_USERS: 25, // в данном случае нужно ли писать заглавными буквами? Или можно и строчными?
+  MIN_OF_LIKES: 15,
+  MAX_OF_LIKES: 200,
+  MAX_OF_COMMENTS: 30,
+  MAX_OF_PHOTOS: 6,
+  MAX_OF_MESSAGES: 2,
+};
+
+export {settings};
